Added new styling for discussion board, Refactored MessageBoard, MessageList components
