+++ components/MessageBoard/MessageList.tsx
@@ -1,16 +1,15 @@
 import { GraphQLResult } from "@aws-amplify/api"
-import { RouteProp, useNavigation, useRoute } from "@react-navigation/native"
+import { RouteProp, useNavigation } from "@react-navigation/native"
 import { StackNavigationProp } from "@react-navigation/stack"
 import { API } from "aws-amplify"
 import GRAPHQL_AUTH_MODE from "aws-amplify-react-native"
 import { convertFromRaw } from "draft-js"
 import { stateToHTML } from "draft-js-export-html"
-import { Body, Card, CardItem, Left, Right } from "native-base"
-import React from "react"
-import { isFirefox, isSafari } from "react-device-detect"
-import { ActivityIndicator, Dimensions, FlatList, Text, TouchableOpacity, View } from "react-native"
-import Observable, { ZenObservable } from "zen-observable-ts"
-import JCComponent, { JCState } from "../../components/JCComponent/JCComponent"
+import moment from "moment"
+import React, { useEffect, useState } from "react"
+import { ActivityIndicator, FlatList, Text, TouchableOpacity, View } from "react-native"
+import Observable from "zen-observable-ts"
+import { JCState } from "../../components/JCComponent/JCComponent"
 import ProfileImage from "../../components/ProfileImage/ProfileImage"
 import {
   GetMessageQuery,
@@ -24,6 +23,7 @@ import {
   onCreateMessageByRoomId,
   onCreateReply,
 } from "../../src/graphql-custom/messages"
+import MessageInput from "./MessageInput"
 import MessageUtils from "./MessageUtils"
 
 type Messages = NonNullable<MessagesByRoomQuery["messagesByRoom"]>["items"]
@@ -42,376 +42,212 @@ interface Props {
 }
 interface State extends JCState {
   nextToken: string | null | undefined
-
   messages: Messages
   fetchingData: boolean
 }
-class MessageListImpl extends JCComponent<Props, State> {
-  flatListRef: React.RefObject<FlatList<any>>
-  messageUnsubscribe?: ZenObservable.Subscription
-  replyUnsubscribe?: ZenObservable.Subscription
 
-  constructor(props: Props) {
-    super(props)
-    this.messageUnsubscribe = undefined
-    this.replyUnsubscribe = undefined
-    this.flatListRef = React.createRef<FlatList<any>>()
-
-    this.state = {
-      ...super.getInitialState(),
-      messages: [],
-      nextToken: null,
-      fetchingData: false,
-    }
-    this.setInitialData()
-  }
-  async setInitialData() {
-    if (this.props.groupId) {
-      this.getMessages()
-    }
-  }
-  componentDidUpdate(prevProps: Props) {
-    if (this.props.groupId !== prevProps.groupId) {
-      this.setInitialData()
-    }
-  }
-
-  componentDidMount() {
-    this.connectSubscriptions()
-    if (this.props.inputAt === "bottom" && !isFirefox && !isSafari) {
-      const scrollNode = this.flatListRef.current && this.flatListRef.current?.getScrollableNode()
-      if (!!scrollNode)
-        scrollNode.addEventListener("wheel", (e: any) => {
-          scrollNode.scrollTop -= e.deltaY
-          e.preventDefault()
-        })
-      this.flatListRef.current?.setNativeProps({
-        style: { transform: "translate3d(0,0,0) scaleY(-1)" },
-      })
-    }
-  }
-  componentWillUnmount() {
-    if (this.props.inputAt === "bottom" && !isFirefox && !isSafari) {
-      const scrollNode = this.flatListRef.current && this.flatListRef.current?.getScrollableNode()
-      if (!!scrollNode)
-        scrollNode.removeEventListener("wheel", (e: any) => {
-          scrollNode.scrollTop -= e.deltaY
-          e.preventDefault()
-        })
-    }
-  }
-  async connectSubscriptions() {
-    if (this.props.groupId) {
-      const messageSub = API.graphql({
-        query: onCreateMessageByRoomId,
-        variables: { roomId: this.props.groupId },
-        authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
-      }) as Observable<{
-        provider: any
-        value: GraphQLResult<OnCreateMessageByRoomIdSubscription>
-      }>
-      this.messageUnsubscribe = messageSub.subscribe({
-        next: (incoming) => {
-          console.debug(incoming)
-          if (incoming.value.data?.onCreateMessageByRoomId) {
-            this.setState({
-              messages: [
-                ...[incoming.value.data.onCreateMessageByRoomId],
-                ...(this.state.messages ?? []),
-              ],
-            })
-          }
-        },
-      })
-      const replySub = API.graphql({
-        query: onCreateReply,
-        authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
-      }) as Observable<{
-        provider: any
-        value: GraphQLResult<OnCreateReplySubscription>
-      }>
-      this.replyUnsubscribe = replySub.subscribe({
-        next: async (incoming) => {
-          console.debug(incoming)
-          if (
-            incoming.value?.data?.onCreateReply?.parentMessage &&
-            incoming.value?.data.onCreateReply?.parentMessage?.roomId === this.props.groupId
-          ) {
-            const { messages } = this.state
-            // find incoming reply's parent message in current state
-            const index = messages?.findIndex(
-              (m) => m?.id === incoming.value?.data?.onCreateReply?.messageId
-            )
-
-            try {
-              const updatedMessage = (await API.graphql({
-                query: getMessage,
-                variables: {
-                  id: incoming.value.data.onCreateReply.messageId,
-                },
-                authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
-              })) as GraphQLResult<GetMessageQuery>
-
-              if (
-                updatedMessage.data?.getMessage &&
-                index !== undefined &&
-                messages &&
-                messages[index]
-              ) {
-                // replace old message/replies with incoming data
-                messages[index] = updatedMessage.data.getMessage
-                this.setState({ messages })
-              }
-            } catch (e: any) {
-              console.debug(e)
-              if (e.data?.getMessage && index !== undefined && messages && messages[index]) {
-                // replace old message/replies with incoming data
-                messages[index] = e.data.getMessage
-                this.setState({ messages })
-              }
-            }
-          }
-        },
-      })
-    }
-  }
-  removeSubscriptions() {
-    if (this.props.groupId) {
-      this.messageUnsubscribe?.unsubscribe()
-      this.replyUnsubscribe?.unsubscribe()
-    }
-  }
-  async getMessages() {
-    try {
-      const messages = (await API.graphql({
-        query: messagesByRoom,
-        variables: {
-          roomId: this.props.groupId,
-          sortDirection: "DESC",
-          limit: 10,
-        },
-        authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
-      })) as GraphQLResult<MessagesByRoomQuery>
-
-      console.debug(messages)
-
-      if (messages.data?.messagesByRoom?.items) {
-        this.setState({
-          messages: messages.data.messagesByRoom.items,
-          nextToken: messages.data.messagesByRoom.nextToken,
-        })
-        this.props.onHandleCreated()
-      }
-    } catch (e: any) {
-      console.debug(e)
-      if (e.data?.messagesByRoom) {
-        this.setState({
-          messages: e.data.messagesByRoom.items,
-          nextToken: e.data.messagesByRoom.nextToken,
-        })
-        this.props.onHandleCreated()
-      }
-    }
-  }
-
-  async getMoreMessages() {
-    this.setState({ fetchingData: true })
-    if (this.state.nextToken) {
+export default function MessageList(props: Props): JSX.Element {
+  const [state, setState] = useState<State>({
+    messages: [],
+    nextToken: null,
+    fetchingData: false,
+  })
+  const getMoreMessages = async () => {
+    setState((prev) => ({ ...prev, fetchingData: true }))
+    if (state.nextToken) {
       try {
         const messages = (await API.graphql({
           query: messagesByRoom,
           variables: {
-            roomId: this.props.groupId,
+            roomId: props.groupId,
             sortDirection: "DESC",
             limit: 10,
-            nextToken: this.state.nextToken,
+            nextToken: state.nextToken,
           },
           authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
         })) as GraphQLResult<MessagesByRoomQuery>
 
         if (messages.data?.messagesByRoom?.items) {
-          this.setState({
-            messages: [...(this.state.messages ?? []), ...messages.data.messagesByRoom.items],
-            nextToken: messages.data.messagesByRoom.nextToken,
-          })
+          setState((prev) => ({
+            ...prev,
+            messages: [...(prev.messages ?? []), ...(messages?.data?.messagesByRoom?.items ?? [])],
+            nextToken: messages?.data?.messagesByRoom?.nextToken,
+          }))
         }
       } catch (e: any) {
         if (e.data?.messagesByRoom) {
-          this.setState({
-            messages: [...(this.state.messages ?? []), ...e.data.messagesByRoom.items],
+          setState((prev) => ({
+            ...prev,
+            messages: [...(prev.messages ?? []), ...e.data.messagesByRoom.items],
             nextToken: e.data.messagesByRoom.nextToken,
-          })
+          }))
         }
       }
     }
-    this.setState({ fetchingData: false })
+    setState((prev) => ({ ...prev, fetchingData: false }))
   }
-  showProfile(id: string | undefined) {
-    if (id) this.props.navigation?.push("ProfileScreen", { id: id, create: false })
-  }
-  renderMessageWithReplies(item: Message, index: number) {
+  const renderMessageWithReplies = (item: Message, index: number) => {
+    //
     return (
-      <View style={{ marginBottom: 20 }} key={index}>
-        {this.renderMessage(item, index, false)}
-        {item?.replies?.items?.map((reply, index) => {
-          return this.renderMessage(reply, index, true)
+      <View
+        style={{
+          marginBottom: 32,
+          borderWidth: 1,
+          borderColor: "#E4E1E1",
+          borderRadius: 8,
+          backgroundColor: "#fff",
+        }}
+        key={index}
+      >
+        {renderMessage(item, index, false)}
+        {item?.replies?.items.slice(0, 2).map((reply, index) => {
+          return renderMessage(reply, index, true)
         })}
+        <TouchableOpacity
+          onPress={() => null} // increase by 2, need to add state for num replies
+        >
+          {item?.replies?.items?.length && item?.replies?.items?.length >= 3 ? (
+            <Text
+              style={{
+                marginLeft: 142,
+                color: "#6A5E5D",
+                fontFamily: "Graphik-Regular-App",
+                fontSize: 14,
+                lineHeight: 24,
+              }}
+            >
+              View {item.replies.items.length - 2} more comments... (out of{" "}
+              {item?.replies?.items?.length})
+            </Text>
+          ) : null}
+        </TouchableOpacity>
+        <MessageInput
+          replyMode={true}
+          replyState={{
+            replyToWho: [],
+            replyToId: item?.id,
+            replyToRoomId: props.groupId ?? "",
+          }}
+          clearReplyState={function (): void {
+            throw new Error("Function not implemented.")
+          }}
+          groupId={props.groupId}
+        />
       </View>
     )
   }
-
-  renderMessage(item: Message | Reply, index: number, isReply: boolean) {
-    const { style, replies } = this.props
+  const renderMessage = (item: Message | Reply, index: number, isReply: boolean) => {
+    const showProfile = (id: string | undefined) => {
+      if (id) navigation?.push("ProfileScreen", { id: id, create: false })
+    }
+    const { replies } = props
+    const now = moment()
+    const datePosted = moment(parseInt(item?.when))
+    const daysSince = now.diff(datePosted.format("L"), "days")
+    const dateString = `${
+      daysSince === 0
+        ? "Today"
+        : daysSince === 1
+        ? "Yesterday"
+        : daysSince > 1
+        ? daysSince + " days ago"
+        : null
+    }`
     return (
-      <Card
+      <View
         key={index}
         style={{
-          borderRadius: 10,
-          minHeight: 50,
-          borderColor: "#ffffff",
-          marginLeft: isReply ? 50 : 0,
-          shadowColor: "#ffffff",
+          paddingHorizontal: 16,
+          paddingVertical: 32,
+          paddingBottom: !isReply ? 32 : 0,
+          marginLeft: isReply ? 64 : 0,
+          borderBottomWidth: !isReply ? 1 : 0,
+          borderBottomColor: "#E4E1E1",
         }}
       >
-        {style === "regular" || style === "course" || style === "courseResponse" ? (
-          <CardItem style={this.styles.style.eventPageMessageBoard}>
-            <Left style={this.styles.style.eventPageMessageBoardLeft}>
+        <View>
+          <View style={{ flexDirection: "row" }}>
+            <View style={{ flexDirection: "column" }}>
               <TouchableOpacity
                 key={item?.id}
                 onPress={() => {
-                  this.showProfile(item?.author?.id)
+                  showProfile(item?.author?.id)
                 }}
               >
                 {item && "owner" in item && (
-                  <ProfileImage size="small2" user={item?.owner ?? null} />
+                  <View style={{ marginRight: 16 }}>
+                    <ProfileImage size="editorLarge" user={item?.owner ?? null} />
+                  </View>
                 )}
-                {isReply && <ProfileImage size="smallReply" user={item?.author?.id ?? null} />}
+                {isReply && (
+                  <View style={{ marginLeft: -16 }}>
+                    <ProfileImage size="small2" user={item?.author?.id ?? null} />
+                  </View>
+                )}
               </TouchableOpacity>
-              <Body>
-                <Text style={this.styles.style.groupFormName}>
-                  {item?.author?.given_name ?? ""} {item?.author?.family_name ?? ""}
-                </Text>
-                <Text style={this.styles.style.groupFormRole}>
-                  {item?.author?.currentRole ?? ""}
-                </Text>
-                {item?.when && (
-                  <Text style={this.styles.style.MessageBoardFormDate}>
-                    {new Date(parseInt(item?.when, 10)).toLocaleDateString()}
-                  </Text>
-                )}
-              </Body>
-            </Left>
-            <Right style={{ justifyContent: "center" }}>
-              {replies && (
-                <TouchableOpacity
-                  style={{
-                    alignSelf: "flex-end",
-                    margin: 24,
-                    borderWidth: 1,
-                    borderColor: "#F0493E",
-                    borderRadius: 4,
-                    paddingTop: 7,
-                    paddingBottom: 7,
-                    paddingLeft: 23,
-                    paddingRight: 23,
-                  }}
-                  onPress={() => this.props.onHandlePressReply(item)}
-                >
+            </View>
+
+            <View style={{ flex: 1 }}>
+              <View style={{ flexDirection: "row", flex: 1 }}>
+                <View style={{ flexDirection: "column", flex: 1 }}>
                   <Text
                     style={{
-                      fontFamily: "Graphik-Regular-App",
-                      fontWeight: "normal",
-                      fontSize: 14,
-                      lineHeight: 20,
-                      color: "#F0493E",
+                      color: "#1A0706",
+                      fontFamily: "Graphik-Semibold-App",
+                      fontSize: 16,
+                      lineHeight: 24,
                     }}
                   >
-                    reply
+                    {item?.author?.given_name ?? ""} {item?.author?.family_name ?? ""}
                   </Text>
-                </TouchableOpacity>
-              )}
-            </Right>
-          </CardItem>
-        ) : (
-          <CardItem style={this.styles.style.coursePageMessageBoard}>
-            <Left style={this.styles.style.coursePageMessageBoardLeftMini}>
-              <TouchableOpacity
-                key={item?.id}
-                onPress={() => {
-                  this.showProfile(item?.author?.id)
-                }}
-              >
-                {item && "owner" in item && (
-                  <ProfileImage size="small2" user={item?.owner ?? null} />
-                )}
-                {isReply && <ProfileImage size="smallReply" user={item?.author?.id ?? null} />}
-              </TouchableOpacity>
-            </Left>
-            <Right style={this.styles.style.miniMessageBoardRight}>
-              <View>
-                <Text style={this.styles.style.courseFormName}>
-                  {item?.author?.given_name ?? ""} {item?.author?.family_name ?? ""}
-                </Text>
-                {item?.when && (
-                  <Text style={this.styles.style.groupFormDate}>
-                    {new Date(parseInt(item?.when, 10)).toLocaleDateString()}
-                  </Text>
-                )}
-              </View>
-              <View>
-                {replies && (
-                  <TouchableOpacity
+                  <Text
                     style={{
-                      alignSelf: "flex-end",
-                      margin: 0,
-                      borderWidth: 1,
-                      borderColor: "#F0493E",
-                      borderRadius: 4,
-                      paddingTop: 4,
-                      paddingBottom: 4,
-                      paddingLeft: 10,
-                      paddingRight: 10,
+                      color: "#483938",
+                      fontFamily: "Graphik-Regular-App",
+                      fontSize: 16,
+                      lineHeight: 24,
                     }}
-                    onPress={() => this.props.onHandlePressReply(item)}
                   >
-                    <Text
-                      style={{
-                        fontFamily: "Graphik-Regular-App",
-                        fontWeight: "normal",
-                        fontSize: 14,
-                        lineHeight: 20,
-                        color: "#F0493E",
-                      }}
-                    >
-                      reply
-                    </Text>
-                  </TouchableOpacity>
+                    {item?.author?.currentRole ?? ""}
+                  </Text>
+                </View>
+                {item?.when && (
+                  <Text
+                    style={{
+                      color: "#6A5E5D",
+                      fontFamily: "Graphik-Regular-App",
+                      fontSize: 16,
+                      lineHeight: 24,
+                    }}
+                  >
+                    {dateString}
+                  </Text>
                 )}
               </View>
-            </Right>
-          </CardItem>
-        )}
-        <CardItem style={this.styles.style.eventPageMessageBoardInnerCard}>
-          <div id="comment-div">
-            <div
-              dangerouslySetInnerHTML={{
-                __html: this.convertCommentFromJSONToHTML(item?.content ?? null),
-              }}
-            ></div>
-          </div>
-        </CardItem>
-        {item?.attachment ? (
-          <CardItem>{MessageUtils.renderFileDownloadBadge(item)}</CardItem>
-        ) : null}
-      </Card>
+
+              <div
+                style={{
+                  fontFamily: "Graphik-Regular-App",
+                  fontWeight: "normal",
+                  fontSize: 15,
+                  lineHeight: "24px",
+                  color: "#1A0706",
+                  marginBlockEnd: 0,
+                }}
+                dangerouslySetInnerHTML={{
+                  __html: convertCommentFromJSONToHTML(item?.content ?? null),
+                }}
+              />
+
+              {item?.attachment ? <View>{MessageUtils.renderFileDownloadBadge(item)}</View> : null}
+            </View>
+          </View>
+        </View>
+      </View>
     )
   }
-
-  convertCommentFromJSONToHTML = (text: string | null) => {
-    const errorMarkdown =
-      "<div>" + this.props.style == "course" || this.props.style == "courseResponse"
-        ? "Assignment/Response"
-        : "Message" + " Can't Be Displayed</div>"
+  const convertCommentFromJSONToHTML = (text: string | null) => {
+    const errorMarkdown = "<div>" + "Message" + " Can't Be Displayed</div>"
 
     if (!text) return errorMarkdown
 
@@ -422,32 +258,141 @@ class MessageListImpl extends JCComponent<Props, State> {
       return errorMarkdown
     }
   }
-
-  messagesLoader() {
-    if (this.state.nextToken) return <ActivityIndicator animating color="#F0493E" />
+  const messagesLoader = () => {
+    if (state.nextToken) return <ActivityIndicator animating color="#F0493E" />
     return null
   }
+  const navigation = useNavigation<StackNavigationProp<any, any>>()
+  const flatListRef = React.createRef<FlatList<any>>()
+  useEffect(() => {
+    const messageSub = (
+      API.graphql({
+        query: onCreateMessageByRoomId,
+        variables: { roomId: props.groupId },
+        authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
+      }) as Observable<{
+        provider: any
+        value: GraphQLResult<OnCreateMessageByRoomIdSubscription>
+      }>
+    ).subscribe({
+      next: (incoming) => {
+        console.log({ incoming })
+        if (incoming.value.data?.onCreateMessageByRoomId) {
+          setState((prev) => ({
+            ...prev,
+            messages: [
+              incoming?.value?.data?.onCreateMessageByRoomId ?? ({} as Message),
+              ...prev.messages,
+            ],
+          }))
+        }
+      },
+    })
+    const replySub = (
+      API.graphql({
+        query: onCreateReply,
+        authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
+      }) as Observable<{
+        provider: any
+        value: GraphQLResult<OnCreateReplySubscription>
+      }>
+    ).subscribe({
+      next: async (incoming) => {
+        console.log({ incoming })
+        if (
+          incoming.value?.data?.onCreateReply?.parentMessage &&
+          incoming.value?.data.onCreateReply?.parentMessage?.roomId === props.groupId
+        ) {
+          const tempMessages = state.messages
+          // find incoming reply's parent message in current state
+          const index = tempMessages?.findIndex(
+            (m) => m?.id === incoming.value?.data?.onCreateReply?.messageId
+          )
+          try {
+            const updatedMessage = (await API.graphql({
+              query: getMessage,
+              variables: {
+                id: incoming.value.data.onCreateReply.messageId,
+              },
+              authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
+            })) as GraphQLResult<GetMessageQuery>
+            console.log({ updatedMessage })
+            console.log({ tempMessages })
 
-  render() {
-    return (
+            console.log("index is ", { index })
+            if (
+              updatedMessage.data?.getMessage &&
+              index !== undefined &&
+              tempMessages &&
+              tempMessages[index]
+            ) {
+              // replace old message/replies with incoming data
+              tempMessages[index] = updatedMessage.data.getMessage
+              setState((prev) => ({ ...prev, messages: tempMessages }))
+            }
+          } catch (e: any) {
+            console.error(e)
+            if (e.data?.getMessage && index !== undefined && tempMessages && tempMessages[index]) {
+              // replace old message/replies with incoming data
+              tempMessages[index] = e.data.getMessage
+              setState((prev) => ({ ...prev, messages: tempMessages }))
+            }
+          }
+        }
+      },
+    })
+
+    const loadMessages = async () => {
+      try {
+        const messages = (await API.graphql({
+          query: messagesByRoom,
+          variables: {
+            roomId: props.groupId,
+            sortDirection: "DESC",
+            limit: 10,
+          },
+          authMode: GRAPHQL_AUTH_MODE.AMAZON_COGNITO_USER_POOLS,
+        })) as GraphQLResult<MessagesByRoomQuery>
+
+        console.debug(messages)
+
+        if (messages.data?.messagesByRoom?.items) {
+          setState((prev) => ({
+            ...prev,
+            messages: messages?.data?.messagesByRoom?.items ?? [],
+            nextToken: messages?.data?.messagesByRoom?.nextToken,
+          }))
+          props.onHandleCreated()
+        }
+      } catch (e: any) {
+        console.debug(e)
+        if (e.data?.messagesByRoom) {
+          setState((prev) => ({
+            ...prev,
+            messages: e.data.messagesByRoom.items,
+            nextToken: e.data.messagesByRoom.nextToken,
+          }))
+          props.onHandleCreated()
+        }
+      }
+    }
+    if (!state.messages.length) loadMessages()
+    return () => {
+      replySub.unsubscribe()
+      messageSub.unsubscribe()
+    }
+  }, [props.groupId, state.messages])
+  return (
+    <View>
       <FlatList
-        ref={this.flatListRef}
+        ref={flatListRef}
         scrollEnabled
-        renderItem={({ item, index }) => this.renderMessageWithReplies(item, index)}
-        data={this.state.messages}
-        inverted={this.props.inputAt === "bottom"}
-        onEndReached={!this.state.fetchingData ? () => this.getMoreMessages() : undefined}
-        style={{ height: Dimensions.get("window").height - 270 }}
-        ListFooterComponent={() => this.messagesLoader()}
-        refreshing={this.state.fetchingData}
+        renderItem={({ item, index }) => renderMessageWithReplies(item, index)}
+        data={state.messages}
+        onEndReached={!state.fetchingData ? () => getMoreMessages() : undefined}
+        refreshing={state.fetchingData}
         onEndReachedThreshold={0.1}
       />
-    )
-  }
-}
-
-export default function MessageList(props: Props): JSX.Element {
-  const route = useRoute()
-  const navigation = useNavigation<StackNavigationProp<any, any>>()
-  return <MessageListImpl {...props} navigation={navigation} route={route} />
+    </View>
+  )
 }
